// core/types/generated/contentful.d.ts
export interface IHeroSlideFields {
  /** Image */
  image?: Asset | undefined;

  /** Headline */
  headline: string;

  /** Subhead */
  subhead?: string | undefined;

  /** CTA Label */
  ctaLabel?: string | undefined;

  /** CTA Link */
  ctaLink?: Entry<{ [fieldId: string]: unknown }> | undefined;
}

export interface IHeroSlide extends Entry<IHeroSlideFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "heroSlide";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPageStandardFields {
  /** Page Name */
  pageName: string;

  /** Meta Title (SEO) */
  metaTitleSeo: string;

  /** Meta Description (SEO) */
  metaDescription?: string | undefined;

  /** Meta Keywords (SEO) */
  metaKeywordsSeo?: string | undefined;

  /** Page Slug */
  pageSlug: string;

  /** Optional Page Description */
  optionalPageDescription?: Document | undefined;

  /** Page Content Entries */
  pageContent?: Entry<{ [fieldId: string]: unknown }>[] | undefined;
}

export interface IPageStandard extends Entry<IPageStandardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pageStandard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}